requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
dvc>=3.38.0
dvc-azure>=3.0.0
azure-storage-blob>=12.19.0
PyYAML>=6.0
evidently>=0.4.0
prometheus_client>=0.20.0
